Update oidc setup

+++ src/deployments/keycloak-oidc/main.sh
@@ -1,24 +1,26 @@
 logger "$INFO" "installing keycloak oidc realm in $namespace namespace...";
 
-sed "s/replace_cr_name/$crName/g" "$deploymentPath/realm-template.yaml" | \
-  sed "s/replace_keycloak/$keycloakName/g" | \
-  sed "s/replace_username/$USERNAME/g" | \
-  sed "s/replace_password/$PASSWORD/g" | \
-  sed "s/replace_secret/$secret/g" | \
-  kubectl apply --namespace "$namespace" -f - > /dev/null;
+checkKeycloakRealmImportCRExists "$crName";
+if ! $keycloakRealmImportCRExists; then
+  sed "s/replace_cr_name/$crName/g" "$keycloakOIDCDeploymentPath/realm-template.yaml" | \
+    sed "s/replace_keycloak/$keycloakName/g" | \
+    sed "s/replace_username/$USERNAME/g" | \
+    sed "s/replace_password/$PASSWORD/g" | \
+    sed "s/replace_secret/$secret/g" | \
+    kubectl apply --namespace "$namespace" -f - > /dev/null;
 
-logger "$DEBUG" "keycloak-oidc: waiting for $crName to be created";
+  logger "$DEBUG" "keycloak-oidc: waiting for $crName to be created";
 
-waitForPods "job-name=$crName" "$crName";
-kubectl wait --namespace "$namespace" jobs "$crName" --for=condition=complete --timeout=120s > /dev/null;
-kubectl delete --namespace "$namespace" jobs "$crName" > /dev/null;
+  waitForPods "job-name=$crName" "$crName";
+  kubectl wait --namespace "$namespace" jobs "$crName" --for=condition=complete --timeout=120s > /dev/null;
 
-logger "$INFO" "waiting for keycloak to be ready...";
-kubectl wait pods --namespace "$namespace" -l "statefulset.kubernetes.io/pod-name=$keycloakName-0" --for=delete --timeout=60s > /dev/null;
-waitForPods "statefulset.kubernetes.io/pod-name=$keycloakName-0" "$keycloakName-0";
-kubectl wait pods --namespace "$namespace" -l "statefulset.kubernetes.io/pod-name=$keycloakName-0" --for=condition=Ready --timeout=180s > /dev/null;
+  logger "$INFO" "waiting for keycloak to be ready...";
+  kubectl wait pods --namespace "$namespace" -l "statefulset.kubernetes.io/pod-name=$keycloakName-0" --for=delete --timeout=60s > /dev/null;
+  waitForPods "statefulset.kubernetes.io/pod-name=$keycloakName-0" "$keycloakName-0";
+  kubectl wait pods --namespace "$namespace" -l "statefulset.kubernetes.io/pod-name=$keycloakName-0" --for=condition=Ready --timeout=180s > /dev/null;
+fi
 
-sed "s/replace_service_url/httpbin-svc.$namespace.svc:8000/g" "$deploymentPath/api-template.yaml" | \
+sed "s/replace_service_url/httpbin-svc.$namespace.svc:8000/g" "$keycloakOIDCDeploymentPath/api-template.yaml" | \
   sed "s/replace_namespace/$namespace/g" | \
   sed "s/replace_realm_url/https:\/\/$keycloakName-service.$namespace.svc:$KEYCLOAK_SERVICE_PORT\/realms\/oidc/g" | \
   sed "s/replace_client_id/$(base64 $client_id)/g" | \
